LayoutTests/svg/animations/animate-points-expected.txt

CONSOLE MESSAGE: line 3: ReferenceError: Can't find variable: createSVGTestCase
SVG 1.1 dynamic animation tests

Tests if points of a polygon are animateable.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
